+++ CHANGES.rst
@@ -8,6 +8,7 @@
 - Watershed: smoothing of result watershed polygon is fixed and configurable via checkbox (#668).
 - Graph: user-defined label in legend is now set per plot instead of per feature (#840)
 - Graph: Use "{grid name} | {result name} | ID {id}" as default label (#840)
+- Graph: Added splitter so plots and legends can be resized (#840)
 
 
 3.1.4 (2023-06-06)